hello world

+++ iikoview.py
@@ -19,28 +19,48 @@ urls = APIRouter()
 
 @urls.get('/synch/department',response_model=Page[schemas.GetFillialSch])
 async def insert_departments(db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)): 
-    data = crud.insert_fillials(db,items=list_departments(key=authiiko()))
-    stores = crud.insert_otdels(db,items=list_stores(key=authiiko()))
-    suppliers = crud.synch_suppliers(db,suppliers=get_suppliers(key=authiiko()))
-    branches = crud.get_branch_list(db)
-    return paginate(branches)
-
+    permission = checkpermissions(request_user=request_user,db=db,page=9)
+    if permission:
+        data = crud.insert_fillials(db,items=list_departments(key=authiiko()))
+        stores = crud.insert_otdels(db,items=list_stores(key=authiiko()))
+        suppliers = crud.synch_suppliers(db,suppliers=get_suppliers(key=authiiko()))
+        branches = crud.get_branch_list(db)
+        return paginate(branches)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 
 @urls.put('/deparment/update')
 async def update_otdel(form_data:schemas.DepartmenUdpate,db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    query = crud.udpatedepartment(db,form_data=form_data)
-    return query
+    permission = checkpermissions(request_user=request_user,db=db,page=23)
+    if permission:
+        query = crud.udpatedepartment(db,form_data=form_data)
+        return query
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 
 
 @urls.get('/synch/groups')
 async def insert_groups(db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    groups = getgroups(key = authiiko())
-    products = getproducts(key=authiiko())
-    group_list = crud.synchtools(db,groups)
-    product_list = crud.synchproducts(db,grouplist=group_list,products=products)
-    return {'success':True}
+    permission = checkpermissions(request_user=request_user,db=db,page=30)
+    if permission:
+        groups = getgroups(key = authiiko())
+        products = getproducts(key=authiiko())
+        group_list = crud.synchtools(db,groups)
+        product_list = crud.synchproducts(db,grouplist=group_list,products=products)
+        return {'success':True}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 
 @urls.get('/tool/iarch',response_model=list[schemas.ToolParentsch])
@@ -55,13 +75,19 @@ async def toolgroups(query:str,db:Session=Depends(get_db),request_user:schemas.U
 
 
 
-
 @urls.post('/v1/expenditure')
 async def insert_expenditure(amount:Annotated[int,Form()],request_id:Annotated[int,Form()],tool_id:Annotated[int,Form()],comment:Annotated[str,Form()]=None,db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
+    permission = checkpermissions(request_user=request_user,db=db,page=28)
+    if permission:
 
-    query_expenditure = crud.addexpenditure(db,request_id=request_id,amount=amount,tool_id=tool_id,user_id=request_user.id,comment=comment)
-    
-    return {'success':True}
+        query_expenditure = crud.addexpenditure(db,request_id=request_id,amount=amount,tool_id=tool_id,user_id=request_user.id,comment=comment)
+        
+        return {'success':True}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 
 @urls.post('/v1/upload/file')
@@ -83,33 +109,59 @@ async def upload_file(request_id:Annotated[int,Form()],files:list[UploadFile]= N
 
 @urls.put('/v1/expanditure/iiko')
 async def synch_expanditure_iiko(form_data:schemas.SynchExanditureiiko,db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    data = crud.check_expanditure_iiko(db,form_data=form_data)
-    for i in data:
-        if i.status==0:
-            query = crud.synch_expanditure_crud(db,id=i.id)
-            send_document_iiko(key=authiiko(),data=query)
-    return True
-
+    permission = checkpermissions(request_user=request_user,db=db,page=26)
+    if permission:
+        data = crud.check_expanditure_iiko(db,form_data=form_data)
+        for i in data:
+            if i.status==0:
+                query = crud.synch_expanditure_crud(db,id=i.id)
+                send_document_iiko(key=authiiko(),data=query)
+        return True
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 @urls.delete('/v1/expanditure')
 async def delete_expanditure(id=int,db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    data = crud.delete_expanditure(db,id)
-    if data:
-        return {'success':True}
+    permission = checkpermissions(request_user=request_user,db=db,page=28)
+    if permission:
+        data = crud.delete_expanditure(db,id)
+        if data:
+            return {'success':True}
+        else:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail="not id not found"
+            )
     else:
         raise HTTPException(
-            status_code=status.HTTP_409_CONFLICT,
-            detail="not id not found"
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
         )
-    
 
 @urls.post('/v1/comments',response_model=schemas.GetComments)
 async def add_comments(form_data:schemas.AddComments,db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    data = crud.add_comment(db=db,form_data=form_data,user_id = request_user.id)
-    return data
+    permission = checkpermissions(request_user=request_user,db=db,page=1)
+    if permission:
+        data = crud.add_comment(db=db,form_data=form_data,user_id = request_user.id)
+        return data
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
 
 @urls.get('/v1/comments',response_model=Page[schemas.GetComments])
-async def add_comments(db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
-    data = crud.get_comment(db=db)
-    return paginate(data)
+async def get_comments(db:Session=Depends(get_db),request_user:schemas.UserFullBack=Depends(get_current_user)):
+    permission = checkpermissions(request_user=request_user,db=db,page=13)
+    if permission:
+        data = crud.get_comment(db=db)
+        return paginate(data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not super user"
+        )
